+++ include/structures.h
@@ -2,6 +2,6 @@
 #define H_STRUCTURES
 
 void placeTreeStructure (short x, uint8_t y, short z);
-void placeSaplingStructure (short x, uint8_t y, short z, uint8_t sapling_block);
+void placeSaplingStructure (short x, uint8_t y, short z, uint16_t sapling_block);
 
 #endif
